all fix resolved

+++ database/migrations/2024_05_09_190053_create_estados_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('cat_estados', function (Blueprint $table) {
             $table->id();
-            $table->bigInteger('pais_id')->unsigned()->index();
+            $table->mediumInteger('pais_id')->unsigned()->index();
             $table->string('nombre', 50);
             $table->string('nombre_corto', 10);
             $table->string('codigo', 10);
